Rename temporary variable to avoid it overwriting a field
"field" was being seen by the class as a class variable because I used
it in a list comprehension to generate field lists.

Now the temporary variable is "x" rather than "field", which is somewhat
less clear but prevents me from having to call del(field) after
generating the lists. Only two lines are affected, so I think the code's
still nice and clear.